pyproject.toml
[tool.poetry]
name = "player-net"
version = "0.1.0"
description = "IA que juega King of Thieves"
authors = ["Gabriel Faundez <gnfaundez@uc.cl>, Nestor Campos <njcampos1@uc.cl>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
PyAutoGUI = "^0.9.53"
tensorflow-gpu = "^2.5.0"
numba = "^0.53.1"
opencv-python = ">=4.4.0.46"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pycodestyle = "^2.7.0"
flake8 = "^3.9.2"
black = {version = "^21.7b0", allow-prereleases = true}
isort = "^5.9.2"
jupyterlab = "^3.0.16"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
